fix mention device-suffix key, bump version, move mentions to lib/, add CI test job
Addresses PR #16 review: full-JID mention inputs with a device suffix
(from reply_to_sender's raw participant JID) now match on the bare
number via jidDecode instead of keeping the suffix in the match key.
Moves mentions.ts out of scripts/ (reference scripts users copy out)
into lib/ (server.ts's runtime imports), since server.ts imports it.
Bumps plugin.json/marketplace.json to 0.15.3 and adds a CI job that
actually runs the test suite.
